% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFun.R
\name{eventFun}
\alias{eventFun}
\title{View the state-wise details.}
\usage{
eventFun(df)
}
\arguments{
\item{df:}{A data frame that contains information of all countries, out of which one would be selected by user on the line graph.}
}
\description{
This function will let the user select the country of preference to display state-wise details of cases/deaths.
}
\author{
Ketan Kabu
}
