% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputFun.R
\name{inputFun}
\alias{inputFun}
\title{Add a drop down set of options for user to select.}
\usage{
inputFun(id, label)
}
\arguments{
\item{id:}{An identification name for input from user, will later be used in server section to generate a plot based of user's choice.}

\item{label:}{What the user sees when asked for input.}
}
\description{
This function will create drop down selection options in app UI for user to visualise a specific country.
}
\author{
Ketan Kabu
}
