% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Just launch the app!}
\usage{
launch_app()
}
\description{
This function launches the shiny app embedded in this package. Use this function to quickly start using the app.
}
