% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafSel.R
\name{leafSel}
\alias{leafSel}
\title{View a leaflet map.}
\usage{
leafSel(country, df, session)
}
\arguments{
\item{country}{A country that the user will select to display country specific visualisation.}

\item{df}{Dataframe that will be used to display a map based on country selected by user.}

\item{session}{Link to shiny app server.}
}
\description{
This function will display a leaflet map for cases/deaths for the country chosen by the user.
}
\author{
Ketan Kabu
}
