% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_final1}
\alias{us_final1}
\title{USA dataset}
\format{
Dataset contains 2008 rows and 7 columns:
\itemize{
\item \strong{date}: Date of recorded cases/deaths.
\item \strong{State_code}: Name of the states within the country for which the covid pandemic was affected.
\item \strong{state}: The number of Indian nationals who are recorded to have been affected.
\item \strong{country}: The number of foreign nationals who are recorded to have been affected with the pandemic.
\item \strong{newCases}: The number of people who have been recovered on that day in that state.
\item \strong{newDeaths}: The number of people who have been recorded for deaths by the virus.
\item \strong{geometry}: Number of cases recorded within the state.
}
}
\usage{
us_final1
}
\description{
Dataset containing cases and deaths through time in USA, till September end.
}
\keyword{datasets}
