% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widthFun.R
\name{widthFun}
\alias{widthFun}
\title{Add width settings.}
\usage{
widthFun(width, box.width, command)
}
\arguments{
\item{width:}{Width setting for rows, personal preference is 6.}

\item{box.width:}{Width setting for box, personal preference is NULL.}

\item{command:}{The expression for UI input}
}
\description{
This function will create a set UI display of width specified by user for 2 levels - row and box.
}
\author{
Ketan Kabu
}
