% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafSel.R
\name{leafSel}
\alias{leafSel}
\title{View a leaflet map.}
\usage{
leafSel(country)
}
\arguments{
\item{country:}{A country that the user will select to display country specific visualisation.}
}
\description{
This function will display a leaflet map for cases/deaths for the country chosen by the user.
}
\author{
Ketan Kabu
}
